requirements.txt
astroid==2.2.5
attrs==19.1.0
awsebcli==3.16.0
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.0
bootstrap4==0.1.0
botocore==1.13.28
bs4==0.0.1
cement==2.8.2
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.4
django-allauth==0.40.0
django-bootstrap4==1.0.1
django-cors-headers==3.2.0
django-extensions==2.2.5
django-filter==2.2.0
django-filters==0.2.1
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
docutils==0.15.2
drf-yasg==1.17.0
entrypoints==0.3
Faker==1.0.7
Flask==1.1.1
future==0.16.0
get==2019.4.13
gunicorn==20.0.4
h5py==2.9.0
heroku==0.1.4
idna==2.7
inflection==0.3.1
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
isort==4.3.21
itsdangerous==1.1.0
itypes==1.1.0
jedi==0.14.1
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
npm==0.1.1
numpy==1.17.0
oauthlib==3.1.0
optional-django==0.1.0
packaging==19.2
pandocfilters==1.4.2
parso==0.5.1
pathspec==0.5.9
pickleshare==0.7.5
Pillow==6.1.0
post==2019.4.13
pprint==0.1
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psycopg2==2.8.4
public==2019.4.13
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.2
pypiwin32==223
pyrsistent==0.15.3
python-dateutil==1.5
python-decouple==3.3
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2019.2
pywin32==227
pywinpty==0.5.5
PyYAML==3.13
pyzmq==18.0.2
qtconsole==4.5.1
query-string==2019.4.13
request==2019.4.13
requests==2.20.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
scipy==1.3.0
selenium==3.141.0
semantic-version==2.5.0
Send2Trash==1.5.0
six==1.11.0
soupsieve==1.9.5
sqlparse==0.3.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.2
tmdbsimple==2.2.0
tornado==6.0.3
traitlets==4.3.2
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.24.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
whitenoise==4.1.4
wrapt==1.11.2
